forgot a long <-> int conversion

+++ src/spring.cpp
@@ -593,7 +593,7 @@ wxString Spring::GetSPScriptTxt( SinglePlayerBattle& battle )
   s += _T("\tGameType=" )+ usync()->GetModArchive(usync()->GetModIndex(battle.GetModName())) + _T(";\n");
   unsigned long uhash;
   battle.GetModHash().ToULong(&uhash);
-  s += wxString::Format( _T("\tModHash=%ld;\n"), (long)uhash );
+  s += wxString::Format( _T("\tModHash=%ld;\n"), (int)uhash );
   wxStringTripleVec optlistEng;
   battle.CustomBattleOptions()->getOptions( &optlistEng, EngineOption );
   for (wxStringTripleVec::iterator it = optlistEng.begin(); it != optlistEng.end(); ++it)
